RDKOSS-105: Enable the ability to configure tag names in SRCREV (#9)
Reason for change:
This class implementation enables the configuration of tag names in SRCREV for Git repositories,
satisfying the developer requirement for RDK. The class method is executed after the recipe is parsed
and the datastore is populated. It updates the SRCREV by obtaining the SHA value of the configured tag name
through a git ls-remote query with the remote server. This implementation assumes that the necessary rules are configured
in Git repositories to ensure the immutability of tags.

Please note that the option to configure tag names in SRCREV is restricted in Yocto version kirkstone
due to the mutable nature of tags in Git and the necessity for bitbake to generate consistent binaries
from the precise source specified in the recipe at any given point in time.

Signed-off-by: mselva006c <mani_selvaraj@comcast.com>
Co-authored-by: mselva006c <mani_selvaraj@comcast.com>