RDKOSS-218: Fix incomplete tag to SHA conversion in linux-meson
Reason for the change: The "linux-meson" recipe fails to trigger the "bb.event.RecipeParsed" event,
causing an incomplete conversion of the tag to SHA in SRCREV. As a result, the build fails.
To resolve this issue, the tag to SHA conversion implementation has been moved to a Python anonymous API.
This API is now automatically called at the end of the recipe parsing process, ensuring a complete conversion.

Signed-off-by: mselva006c <mani_selvaraj@comcast.com>